requirements.txt
streamlit
openai
langchain-core
langchain-openai
langgraph
pymatgen
ase
crystal-toolkit
pydantic==2.10.3
pytest
python-dotenv